Improve LightSchedule time parsing and organize
- This prepares for adding compatibility with time.Kitchen from the
  frontend but first I need to handle timezone
