+++ garden-app/pkg/garden.go
@@ -2,78 +2,15 @@ package pkg
 
 import (
 	"context"
-	"encoding/json"
 	"errors"
 	"fmt"
 	"net/http"
 	"regexp"
-	"strings"
 	"time"
 
 	"github.com/calvinmclean/automated-garden/garden-app/pkg/influxdb"
 	"github.com/calvinmclean/babyapi"
 )
-
-const (
-	// LightTimeFormat is used to control format of time fields
-	LightTimeFormat = "15:04:05-07:00"
-)
-
-const (
-	// LightStateOff is the value used to turn off a light
-	LightStateOff LightState = iota
-	// LightStateOn is the value used to turn on a light
-	LightStateOn
-	// LightStateToggle is the empty value that results in toggling
-	LightStateToggle
-)
-
-var (
-	stateToString = []string{"OFF", "ON", ""}
-	stringToState = map[string]LightState{
-		`"OFF"`: LightStateOff,
-		`OFF`:   LightStateOff,
-		`"ON"`:  LightStateOn,
-		`ON`:    LightStateOn,
-		`""`:    LightStateToggle,
-		``:      LightStateToggle,
-	}
-)
-
-// LightState is an enum representing the state of a Light (ON or OFF)
-type LightState int
-
-// Return the string representation of this LightState
-func (l LightState) String() string {
-	return stateToString[l]
-}
-
-// MarshalJSON will convert LightState into it's JSON string representation
-func (l LightState) MarshalJSON() ([]byte, error) {
-	if int(l) >= len(stateToString) {
-		return []byte{}, fmt.Errorf("cannot convert %d to %T", int(l), l)
-	}
-	return json.Marshal(stateToString[l])
-}
-
-// UnmarshalJSON with convert LightState's JSON string representation, ignoring case, into a LightState
-func (l *LightState) UnmarshalJSON(data []byte) error {
-	return l.unmarshal(data)
-}
-
-func (l *LightState) UnmarshalText(data []byte) error {
-	return l.unmarshal(data)
-}
-
-func (l *LightState) unmarshal(data []byte) error {
-	upper := strings.ToUpper(string(data))
-	var ok bool
-	*l, ok = stringToState[upper]
-	if !ok {
-		return fmt.Errorf("cannot unmarshal %s into Go value of type %T", string(data), l)
-	}
-	return nil
-}
 
 // Garden is the representation of a single garden-controller device
 type Garden struct {
@@ -101,32 +38,6 @@ type GardenHealth struct {
 	Status      string     `json:"status,omitempty"`
 	Details     string     `json:"details,omitempty"`
 	LastContact *time.Time `json:"last_contact,omitempty"`
-}
-
-// LightSchedule allows the user to control when the Garden light is turned on and off
-// "Time" should be in the format of LightTimeFormat constant ("15:04:05-07:00")
-type LightSchedule struct {
-	Duration    *Duration  `json:"duration" yaml:"duration"`
-	StartTime   string     `json:"start_time" yaml:"start_time"`
-	AdhocOnTime *time.Time `json:"adhoc_on_time,omitempty" yaml:"adhoc_on_time,omitempty"`
-}
-
-// String...
-func (ls *LightSchedule) String() string {
-	return fmt.Sprintf("%+v", *ls)
-}
-
-// Patch allows modifying the struct in-place with values from a different instance
-func (ls *LightSchedule) Patch(new *LightSchedule) {
-	if new.Duration != nil {
-		ls.Duration = new.Duration
-	}
-	if new.StartTime != "" {
-		ls.StartTime = new.StartTime
-	}
-	if new.AdhocOnTime == nil {
-		ls.AdhocOnTime = nil
-	}
 }
 
 // Health returns a GardenHealth struct after querying InfluxDB for the Garden controller's last contact time
@@ -265,9 +176,9 @@ func (g *Garden) Bind(r *http.Request) error {
 				return errors.New("missing required light_schedule.start_time field")
 			}
 			// Check that LightSchedule.StartTime is valid
-			_, err := time.Parse(LightTimeFormat, g.LightSchedule.StartTime)
+			_, err := g.LightSchedule.ParseStartTime()
 			if err != nil {
-				return fmt.Errorf("invalid time format for light_schedule.start_time: %s", g.LightSchedule.StartTime)
+				return err
 			}
 		}
 	case http.MethodPatch:
@@ -291,9 +202,9 @@ func (g *Garden) Bind(r *http.Request) error {
 			}
 			// Check that LightSchedule.StartTime is valid
 			if g.LightSchedule.StartTime != "" {
-				_, err := time.Parse(LightTimeFormat, g.LightSchedule.StartTime)
+				_, err := g.LightSchedule.ParseStartTime()
 				if err != nil {
-					return fmt.Errorf("invalid time format for light_schedule.start_time: %s", g.LightSchedule.StartTime)
+					return err
 				}
 			}
 		}
